Moved PeerError from gather to core and GameEstablished from core to game packages.
